Move TestAdapter tests into ember-testing
